Moved app definition to another file
Also added a DB_TEST variable in the Settings.
