+++ app/main.py
@@ -1,54 +1,14 @@
 import sys
 import os
-
-from fastapi.middleware.cors import CORSMiddleware
-
 sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), "..")))
-
-from fastapi import FastAPI
 import uvicorn
-from motor.motor_asyncio import AsyncIOMotorClient
 from config import settings
+from app_definition import app
 
-from app.routers import user_router
-from app.routers import recipe_router
-
-app = FastAPI()
-
-origins = [
-    "http://localhost:3000",
-    "http://127.0.0.1:3000",
-    "http://0.0.0.0:8000",
-    "http://127.0.0.1:8000"
-]
-
-app.add_middleware(
-    CORSMiddleware,
-    allow_origins=origins,
-    allow_credentials=True,
-    allow_methods=["*"],
-    allow_headers=["*"],
-)
-
-print("\nENVIRONMENT VARIABLES")
-print("DEBUG_MODE: " + str(settings.DEBUG_MODE))
-print("DB_URL: " + settings.DB_URL)
-print("DB_NAME: " + settings.DB_NAME + "\n")
-
-@app.on_event("startup")
-async def startup_db_client():
-    app.mongodb_client = AsyncIOMotorClient(settings.DB_URL)
-    app.mongodb = app.mongodb_client[settings.DB_NAME]
-
-
-@app.on_event("shutdown")
-async def shutdown_db_client():
-    app.mongodb_client.close()
-
-
-app.include_router(user_router.router, tags=["users"], prefix="/user")
-app.include_router(recipe_router.router, tags=["recipes"], prefix="/recipe")
-
+# print("\nENVIRONMENT VARIABLES")
+# print("DEBUG_MODE: " + str(settings.DEBUG_MODE))
+# print("DB_URL: " + settings.DB_URL)
+# print("DB_NAME: " + settings.DB_NAME + "\n")
 
 if __name__ == "__main__":
     uvicorn.run(
@@ -57,7 +17,3 @@ if __name__ == "__main__":
         reload=settings.DEBUG_MODE,
         port=settings.PORT,
     )
-
-@app.get("/")
-async def read_main():
-    return {"msg": "Hello World"}
